refactor: change file according to @ghiggi wishes
