+++ disdrodb/L0/template_tools.py
