+++ disdrodb/L0/templates/NCAR/temp_parser_MIPS.py
@@ -39,7 +39,7 @@ from disdrodb.L0.io import (
     create_directory_structure,
 )
 # Tools to develop the parser  
-from disdrodb.L0.dev_tools import (
+from disdrodb.L0.template_tools import (
     check_column_names,
     infer_df_str_column_names, 
     print_df_first_n_rows,
